CosmosDb provider support for a current region option
Unit and end-to-end tests for the CosmosDbOptionsExtension.WithRegion configuration method

Fix aspnet#15007
